requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.8.5
pandas==2.0.3
matplotlib==3.7.2
qrcode==7.4.2
Pillow==10.0.0
cryptography==41.0.3
psutil==5.9.5
schedule==1.2.0 